+++ plugins/propclass/quest/questfact.h
@@ -63,7 +63,7 @@ private:
     propid_state
   };
   static Property* properties;
-  static int propertycount;
+  static size_t propertycount;
   static void UpdateProperties (iObjectRegistry* object_reg);
 
   // Other fields.
